pyproject.toml
[tool.poetry]
name = "bbdc-bot"
version = "0.1.0"
description = ""
authors = ["xplus2g4 <xplus2g4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.3"
beautifulsoup4 = "^4.11.1"
parse = "^1.19.0"
PyYAML = "^6.0"


[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
